requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
convertdate==2.4.0
    # via holidays
hijri-converter==2.2.4
    # via holidays
holidays==0.18
    # via -r requirements.in
korean-lunar-calendar==0.3.1
    # via holidays
pymeeus==0.5.12
    # via
    #   convertdate
    #   holidays
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   holidays
pytz==2022.7.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
